+++ packages/relation/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## 1.1.0
+## 2.0.0
 
-* Update `rxdart` dependency to `0.27.0`
+* **Breaking Change:** `EntityStateBuilder`'s `child` argument replaced with `builder`
+* **Breaking Change:** `EntityStateBuilder`'s `errorBuilder` argument now passes `Widget Function(BuildContext, Exception)`.
+If you'd like to get `data` from `error` please use `errorDataBuilder` instead.
+* Add the ability to get an `error` in `EntityStateBuilder` from `errorChild`.
+* Add `data` argument to `EntityStreamedState`.
+* Update `rxdart` dependency to `0.27.0`.
 
 ## 1.0.0